ya
ya

+++ src/logica/operacion.java
@@ -2,35 +2,35 @@ package logica;
 
 public class Operacion {
 	
-	private float num1;
-	private float num2;
+	private double num1;
+	private double num2;
 	private char operador;
 	
-	public Operacion(float num1, float num2, char operador) {
+	public Operacion(double num1, double num2, char operador) {
 		super();
 		this.num1 = num1;
 		this.num2 = num2;
 		this.operador = operador;
 	}
 	
-	private float suma(){
+	private double suma(){
 		return num1+num2;
 	}
 	
-	private float resta(){
+	private double resta(){
 		return num1-num2;
 	}
 	
-	private float multiplicacion(){
+	private double multiplicacion(){
 		return num1*num2;
 	}
 	
-	private float division(){
+	private double division(){
 		return num1/num2;
 	}
 	
-	public float controlador(){
-		float result=0;
+	public double controlador(){
+		double result=0;
 		if(operador=='+')
 			result=suma();
 		else if(operador=='-')
